+++ packages/blade/src/components/TopNav/TabNav/index.ts
@@ -1,2 +1,3 @@
 export * from './TabNav';
 export * from './TabNavItem';
+export * from './types';
