+++ packages/blade/src/components/TopNav/index.ts
@@ -1,0 +1,2 @@
+export * from './TabNav';
+export * from './TopNav';
