+++ packages/blade/src/tokens/global/size.ts
@@ -62,6 +62,8 @@ export const size = {
   64: 64,
   /** 72 px */
   72: 72,
+  /** 80 px */
+  80: 80,
   /** 100 px */
   100: 100,
   /** 120 px */
